Upload de img lateral feita, falta integrar

+++ classes/Marca.class.php
@@ -1,23 +1,43 @@
 <?php
 
 include_once "configs/funcoes.php";
+include_once "Imagem.class.php";
 
 class Marca
 {
 	function Marca()
     {
 		$this->entidade = "marca";
+
+		$this->pathProjeto = PATH_SERVIDOR;
+		
+		$classImagem = new Imagem();
+		$this->classImagem = $classImagem;
     }
 	
 	function Grava($post, $file, $copy = NULL)
 	{
 		$retorno = array();
+
+		//Checa se existe imagem
+		if ($file['name'] != "")
+		{
+			//Grava a Imagem
+			$retornoClassImagem = $this->classImagem->gravaImagem($file, $post['extencoeValidas'], "upload/marcas/", $copy);
+			if( $retornoClassImagem[0] )
+			{
+				$retorno[0] = "1";
+				$retorno[1] = $retornoClassImagem[1];
+				return $retorno;
+			}
+		}
 		
 		$sql = "
 			INSERT INTO
 				".$this->entidade."
 			SET
-				titulo 			= '".utf8_decode(str_replace("'","''", $post['titulo']))."'
+				titulo 			= '".utf8_decode(str_replace("'","''", $post['titulo']))."',
+				caminhoImagem	= '".$retornoClassImagem[1]."'
 		";
 		$result = mysql_query($sql);
 		if (!($result))
@@ -37,12 +57,46 @@ class Marca
 	function Altera($post, $file, $copy = NULL)
 	{
 		$retorno = array();
+
+		//Checa se será autializado a imagem
+		if ($file['name'] != "")
+		{
+			if( $post["caminhoImagem"] != "" )
+			{
+				//Checa se existe a imagem
+				if( file_exists($post["caminhoImagem"]) )
+				{
+					//Exclui a antiga
+					if(!unlink( $this->pathProjeto.$post["caminhoImagem"] ))
+					{
+						$retorno[0] = "1";
+						$retorno[1] = "Não foi possíel excluir a imagem antiga!";
+						return $retorno;
+					}
+				}
+			}
+			
+			//Grava a Imagem
+			$retornoClassImagem = $this->classImagem->gravaImagem($file, $post['extencoeValidas'], "upload/marcas/", $copy);
+			if( $retornoClassImagem[0] )
+			{
+				$retorno[0] = "1";
+				$retorno[1] = $retornoClassImagem[1];
+				return $retorno;
+			}
+		}
+		else
+		{
+			//Mantém a imagem antiga
+			$retornoClassImagem[1] = $post["caminhoImagem"];
+		}
 	
 		$sql = "
 			UPDATE	
 				".$this->entidade."
 			SET
-				titulo 			= '".utf8_decode(str_replace("'","''", $post['titulo']))."'
+				titulo 			= '".utf8_decode(str_replace("'","''", $post['titulo']))."',
+				caminhoImagem	= '".$retornoClassImagem[1]."'
 			WHERE
 				id 				= '".$post['id']."'
 		".$query;
@@ -104,6 +158,17 @@ class Marca
 	function Exclui($id)
 	{
 		$retorno = array();
+
+		// Localiza imagem para exclusão do registro
+		$parametro['id'] = $id;
+		$dados = $this->Pesquisar($parametro, null, null);
+		if( $dados[0] )
+		{
+			$retorno[0] = "1";
+			$retorno[1] = "Não foi possível localizar a imagem para excluir!";
+			return $retorno;
+		}
+		$imgExclusao = $dados[1][0]["caminhoImagem"];
 		
 		$sql = "
 			DELETE FROM	
@@ -117,6 +182,17 @@ class Marca
 			$retorno[0] = "1";
 			$retorno[1] = "Erro ao executar a query. Classe = ".$this->entidade." - Metodo = Exclui";
 			return $retorno;
+		}
+
+		//Exclui imagem da pasta
+		if($imgExclusao != "")
+		{ 
+			if(!unlink( $this->pathProjeto.$imgExclusao ))
+			{
+				$retorno[0] = "1";
+				$retorno[1] = "Não foi possíel excluir a imagem do registro!";
+				return $retorno;
+			}
 		}
 		
 		$retorno[0] = 0;
